feat(index.js): 增加日志输出，完善部分逻辑

+++ index.js
@@ -3,15 +3,15 @@ const glob = require("glob");
 const Client = require("ssh2-sftp-client");
 const chalk = require("chalk");
 
-const redLog = function() {
+const redLog = function () {
   console.log.call(null, chalk.red([...arguments]));
 };
 
-const greenLog = function() {
+const greenLog = function () {
   console.log.call(null, chalk.green([...arguments]));
 };
 
-const blueLog = function() {
+const blueLog = function () {
   console.log.call(null, chalk.blue([...arguments]));
 };
 
@@ -27,20 +27,20 @@ const sftp = new Client();
  * 路径结尾要添加  '/'
  */
 class SftpPlugin {
-  // dir 本地目录格式 path.join(__dirname, '..', 'dist/')
+  // localDir 本地目录格式 path.join(__dirname, '..', 'dist/')
 
   constructor({
-    dir = "./dist/",
-    url = "/www/",
+    localDir = "./dist/",
+    remoteDir = "/www/",
     host = "192.168.0.1",
     port = "22",
     username = "username",
     password = "password",
     readyTimeout = 20000,
-    filterFile = null
+    filterFile = null,
   } = {}) {
-    this.url = url;
-    this.dir = dir;
+    this.remoteDir = remoteDir;
+    this.localDir = localDir;
     this.filterFile = filterFile;
     this.startTime = null;
     this.endTime = null;
@@ -49,7 +49,7 @@ class SftpPlugin {
       port,
       username,
       password,
-      readyTimeout
+      readyTimeout,
     };
   }
 
@@ -67,7 +67,7 @@ class SftpPlugin {
 
   put() {
     // 自动上传到FTP服务器
-    if (!this.dir) {
+    if (!this.localDir) {
       redLog("无法上传SFTP,请检查参数");
       return;
     }
@@ -78,10 +78,11 @@ class SftpPlugin {
         // 连接服务器
         blueLog(`连接成功...`);
         sftp
-          .list(this.url)
+          .list(this.remoteDir)
           .then(list => {
-            blueLog(`正在清理文件...`);
+            redLog(`正在清理文件...`);
             this.deleteServerFile(list).then(() => {
+              blueLog(`开始上传文件...`);
               this.globLocalFile();
             });
           })
@@ -95,32 +96,30 @@ class SftpPlugin {
   }
 
   async deleteServerFile(list) {
+    if (list.length === 0) return;
     // 删除服务器上文件(夹)
     for (const fileInfo of list) {
-      const path = this.url + fileInfo.name;
-
+      const path = this.remoteDir + fileInfo.name;
+      redLog(`deleting file: ${path}`);
       if (fileInfo.type === "-") {
         await sftp.delete(path);
       } else {
         await sftp.rmdir(path, true);
       }
     }
-
-    return new Promise(resovle => {
-      resovle();
-    });
   }
 
   globLocalFile() {
     // 获取本地路径所有文件
-
-    glob(this.dir + "**", (er, files) => {
+    const folder = this.localDir + `**`;
+    glob(folder, (er, files) => {
       // 本地目录下所有文件(夹)的路径
 
       files.splice(0, 1); // 删除路径../dist/
 
-      if (this.filterFile && typeof this.filterFile === "function")
+      if (this.filterFile && typeof this.filterFile === "function") {
         files = files.filter(x => this.filterFile(x));
+      }
 
       this.uploadFileToSftp(files);
     });
@@ -130,8 +129,8 @@ class SftpPlugin {
     // 传输文件到服务器
     for (const localSrc of files) {
       const targetSrc = localSrc.replace(
-        this.dir.replace(/\\/g, "/"),
-        this.url
+        this.localDir.replace(/\\/g, "/"),
+        this.remoteDir
       );
 
       if (fs.lstatSync(localSrc).isDirectory()) {
